Update youtubei.js+13.3.0.patch
note that this fix uses the last successfully cached ytb client from github actions. it does NOT check if the client is relevant.

yt.js will fail to cache if sic or nsic failed to extract. nsic is probably the more important one but sic seems to only affect speed, which is irrelevant for music purposes. i've patched the netlify solution to cache despite a failed sic extraction. thus without this patch is also fine; i can also patch the ghactions client to force cache instances with failed sic extraction as well.